feat: Fixing ts files to jsx
