+++ src/store/index.js
@@ -1,0 +1,17 @@
+import { createStore, applyMiddleware, compose } from "redux";
+import createSagaMiddleware from "redux-saga";
+import rootReducer from "./modules/rootReducer";
+import rootSaga from "./modules/rootSaga";
+
+const sagaMiddleware = createSagaMiddleware();
+
+const composeEnhancers = compose;
+
+const store = createStore(
+  rootReducer,
+  composeEnhancers(applyMiddleware(sagaMiddleware))
+);
+
+sagaMiddleware.run(rootSaga);
+
+export default store;
